Split tests and use HTTPS to talk with S3
