+++ composer.json
@@ -15,7 +15,7 @@
         "ext-curl": "*"
     },
     "autoload": {
-        "classmap": ["S3.php"]
+        "classmap": ["S3.php", "tests/S3BaseTest.php"]
     },
     "require-dev": {
         "phpunit/phpunit": "9.x",
